requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/626bfa9c-8221-11ef-b9bf-daac7d5d70b1/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.8.1
awscli==1.36.0
awsebcli==3.21.0
blessed==1.20.0
botocore==1.35.59
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==4.2.16
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
docutils==0.16
future @ file:///AppleInternal/Library/BuildRoots/626bfa9c-8221-11ef-b9bf-daac7d5d70b1/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
idna==3.10
jmespath==1.0.1
macholib @ file:///AppleInternal/Library/BuildRoots/626bfa9c-8221-11ef-b9bf-daac7d5d70b1/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
mysql==0.0.3
mysqlclient==2.2.6
pathspec==0.10.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.3
semantic-version==2.10.0
six @ file:///AppleInternal/Library/BuildRoots/626bfa9c-8221-11ef-b9bf-daac7d5d70b1/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.5.1
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
